TravelAssistant/api_gateway/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
httpx==0.24.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.3.0
pydantic-settings==2.0.3 